+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: tlakchai <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/28 12:47:40 by tlakchai          #+#    #+#             */
-/*   Updated: 2023/08/30 11:29:41 by tlakchai         ###   ########.fr       */
+/*   Updated: 2023/08/30 18:07:35 by tlakchai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	char			*result;
-	unsigned int	i;
+	char	*result;
+	size_t	i;
 
 	if (!s)
 		return (NULL);
